requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
python-dateutil==2.8.2

